krishi-connect-backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
mysqlclient
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart
alembic
python-dotenv
google-genai
mysql-connector-python
